DRTII-1608 Version api. Fix forward ref
