+++ server/src/test/scala/services/exports/QueueExportSpec.scala
@@ -5,7 +5,8 @@ import akka.stream.Materializer
 import drt.shared.CrunchApi.CrunchMinute
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
-import services.exports.QueueExport.{PeriodJson, PortQueuesJson, QueueJson, TerminalQueuesJson}
+import services.api.QueueExport
+import services.api.QueueExport.{PeriodJson, PortQueuesJson, QueueJson, TerminalQueuesJson}
 import uk.gov.homeoffice.drt.ports.PortCode
 import uk.gov.homeoffice.drt.ports.Queues.{EGate, EeaDesk, NonEeaDesk}
 import uk.gov.homeoffice.drt.ports.Terminals.{T1, Terminal}
